+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Bug Fixes
 * Add Index Mapping Update When Node Starts and detect old index mapping. ([#318](https://github.com/opensearch-project/search-relevance/pull/318))
+* Support for handling index mapping differences during node startup. ([#318](https://github.com/opensearch-project/search-relevance/pull/318))
 
 ### Infrastructure